refactor(snapshot): unify ancestors
